# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Bluesky Feed Consumer",
        "description": "Real-time Bluesky firehose consumer with rolling statistics and AI persona chat. Stats are computed in configurable time windows and streamed live via SSE. Persona chat lets you have conversations with AI impersonations of Bluesky accounts, powered by Claude.",
        "version": "0.1.0"
    },
    "paths": {
        "/stats/{window}": {
            "get": {
                "tags": [
                    "stats"
                ],
                "summary": "Get Stats",
                "description": "Return current metrics for a specific time window.\n\nWindow is specified in seconds (e.g., 60 for 1-minute, 300 for 5-minute).\nReturns post/like/repost counts, top-N lists, and language breakdown.",
                "operationId": "get_stats_stats__window__get",
                "parameters": [
                    {
                        "name": "window",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Window"
                        }
                    },
                    {
                        "name": "top_n",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 50,
                            "minimum": 1,
                            "description": "Number of top items to return (max 50)",
                            "default": 10,
                            "title": "Top N"
                        },
                        "description": "Number of top items to return (max 50)"
                    },
                    {
                        "name": "x-api-key",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "X-Api-Key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": true,
                                    "title": "Response Get Stats Stats  Window  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/personas": {
            "post": {
                "tags": [
                    "personas"
                ],
                "summary": "Create Persona",
                "description": "Register a new persona for AI chat.",
                "operationId": "create_persona_personas_post",
                "parameters": [
                    {
                        "name": "x-api-key",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "X-Api-Key"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreatePersonaRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PersonaResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "personas"
                ],
                "summary": "List Personas",
                "description": "List all registered personas.",
                "operationId": "list_personas_personas_get",
                "parameters": [
                    {
                        "name": "x-api-key",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "X-Api-Key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PersonaResponse"
                                    },
                                    "title": "Response List Personas Personas Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/personas/{handle}/status": {
            "get": {
                "tags": [
                    "personas"
                ],
                "summary": "Get Persona Status",
                "description": "Get detailed status for a persona.",
                "operationId": "get_persona_status_personas__handle__status_get",
                "parameters": [
                    {
                        "name": "handle",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Handle"
                        }
                    },
                    {
                        "name": "x-api-key",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "X-Api-Key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PersonaStatusResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/personas/{handle}/chat": {
            "post": {
                "tags": [
                    "personas"
                ],
                "summary": "Post Chat Message",
                "description": "Send a message and receive a streaming response via SSE.",
                "operationId": "post_chat_message_personas__handle__chat_post",
                "parameters": [
                    {
                        "name": "handle",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Handle"
                        }
                    },
                    {
                        "name": "x-api-key",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "X-Api-Key"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChatMessageRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "personas"
                ],
                "summary": "Get Chat History",
                "description": "Retrieve conversation history for a persona.",
                "operationId": "get_chat_history_personas__handle__chat_get",
                "parameters": [
                    {
                        "name": "handle",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Handle"
                        }
                    },
                    {
                        "name": "x-api-key",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "X-Api-Key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChatHistoryResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "personas"
                ],
                "summary": "Delete Chat History",
                "description": "Clear conversation history for a persona.",
                "operationId": "delete_chat_history_personas__handle__chat_delete",
                "parameters": [
                    {
                        "name": "handle",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Handle"
                        }
                    },
                    {
                        "name": "x-api-key",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "X-Api-Key"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/sse/stats": {
            "get": {
                "tags": [
                    "sse"
                ],
                "summary": "Stream Stats",
                "description": "Live stats stream via SSE.\n\nSends a full snapshot on connect, then pushes updates every 2 seconds.\nEach update rotates the velocity bucket and includes current window stats.",
                "operationId": "stream_stats_sse_stats_get",
                "parameters": [
                    {
                        "name": "x-api-key",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "X-Api-Key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "tags": [
                    "system"
                ],
                "summary": "Health",
                "description": "Service health check.",
                "operationId": "health_health_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "string"
                                    },
                                    "type": "object",
                                    "title": "Response Health Health Get"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ChatHistoryResponse": {
                "properties": {
                    "handle": {
                        "type": "string",
                        "title": "Handle"
                    },
                    "messages": {
                        "items": {
                            "$ref": "#/components/schemas/ChatMessageResponse"
                        },
                        "type": "array",
                        "title": "Messages"
                    }
                },
                "type": "object",
                "required": [
                    "handle",
                    "messages"
                ],
                "title": "ChatHistoryResponse"
            },
            "ChatMessageRequest": {
                "properties": {
                    "message": {
                        "type": "string",
                        "title": "Message"
                    }
                },
                "type": "object",
                "required": [
                    "message"
                ],
                "title": "ChatMessageRequest",
                "examples": [
                    {
                        "message": "Hey! What have you been up to lately?"
                    }
                ]
            },
            "ChatMessageResponse": {
                "properties": {
                    "role": {
                        "type": "string",
                        "title": "Role"
                    },
                    "content": {
                        "type": "string",
                        "title": "Content"
                    },
                    "created_at": {
                        "type": "string",
                        "title": "Created At"
                    }
                },
                "type": "object",
                "required": [
                    "role",
                    "content",
                    "created_at"
                ],
                "title": "ChatMessageResponse"
            },
            "CreatePersonaRequest": {
                "properties": {
                    "handle": {
                        "type": "string",
                        "title": "Handle"
                    }
                },
                "type": "object",
                "required": [
                    "handle"
                ],
                "title": "CreatePersonaRequest",
                "examples": [
                    {
                        "handle": "example.bsky.social"
                    }
                ]
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "PersonaResponse": {
                "properties": {
                    "handle": {
                        "type": "string",
                        "title": "Handle"
                    },
                    "display_name": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Display Name"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "post_count": {
                        "type": "integer",
                        "title": "Post Count"
                    },
                    "reply_count": {
                        "type": "integer",
                        "title": "Reply Count"
                    },
                    "last_corpus_update": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Last Corpus Update"
                    },
                    "avatar_url": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Avatar Url"
                    }
                },
                "type": "object",
                "required": [
                    "handle",
                    "display_name",
                    "status",
                    "post_count",
                    "reply_count",
                    "last_corpus_update",
                    "avatar_url"
                ],
                "title": "PersonaResponse"
            },
            "PersonaStatusResponse": {
                "properties": {
                    "handle": {
                        "type": "string",
                        "title": "Handle"
                    },
                    "display_name": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Display Name"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "post_count": {
                        "type": "integer",
                        "title": "Post Count"
                    },
                    "reply_count": {
                        "type": "integer",
                        "title": "Reply Count"
                    },
                    "last_corpus_update": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Last Corpus Update"
                    },
                    "avatar_url": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Avatar Url"
                    },
                    "created_at": {
                        "type": "string",
                        "title": "Created At"
                    },
                    "error_message": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Error Message"
                    }
                },
                "type": "object",
                "required": [
                    "handle",
                    "display_name",
                    "status",
                    "post_count",
                    "reply_count",
                    "last_corpus_update",
                    "avatar_url",
                    "created_at",
                    "error_message"
                ],
                "title": "PersonaStatusResponse"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}